Remove unnecessary scenarios
